Enable hermetic builds for Thanos
This commit enables hermetic builds for Thanos. It also modifies its
Dockerfile to skip using promu and removes the rpm fetcher configuration
given that COO is in the public cluster and can't access internal RH
network.

This is dependent on https://github.com/openshift/thanos/pull/159.
Otherwise this would fail to find the golang toolchain.

Signed-off-by: Daniel Mellado <dmellado@fedoraproject.org>